cpufreq: added Optimax CPU GOV
This is based on ONDEMAND, like almost all governors that have arisen from XDA.
It contains some enhancements from LG, particularly to freq boost handling so
it will boost to a set level, almost like HTC's governor. It has different
tunables to the HTC governor but it behaves pretty similar, the tunables it
comes with default are a bit more conservative.
It originates from Cl3kener's Uber kernel for Nexus 5, where it has quite a
reputation for battery life

Based on commits:
 https://github.com/VanirAOSP/kernel_htc_msm8974/commit/ca78ca2c367d07da5cbf4f3b7adf3a251e6c7400
 https://github.com/VanirAOSP/kernel_htc_msm8974/commit/665007f0b55ef40c2809e05382ab465995176e35
 https://github.com/VanirAOSP/kernel_htc_msm8974/commit/6c623a9de704d2e66b0d0ceb1b534d28ee09b042
 https://github.com/VanirAOSP/kernel_htc_msm8974/commit/b7a05d66ea7abbdae505089e4875cf5f3abbd287