cpufreq: added DanceDance CPU GOV
Based on conservative with some smartass features, it scales accordingly
to conservatives laws. So it will start from the bottom, take a load sample,
if it's above the upthreshold, ramp up only one speed at a time, and ramp
down one at a time. It will automatically cap the off screen speeds to
245Mhz, and if your min freq is higher than 245mhz, it will reset the min
to 120mhz while screen is off and restore it upon screen awakening, and
still scale accordingly to conservatives laws. So it spends most of its
time at lower frequencies. The goal of this is to get the best battery
life with decent performance. It is a performance focused governor but
also blends with some battery savings

Based on commits:
 https://github.com/VanirAOSP/kernel_htc_msm8974/commit/ca78ca2c367d07da5cbf4f3b7adf3a251e6c7400
 https://github.com/VanirAOSP/kernel_htc_msm8974/commit/9796346987465066163fe65d9319cab31e1c0703
 https://github.com/VanirAOSP/kernel_htc_msm8974/commit/767475c19c6b3b28c2c8579db38bb4e004e94dd0
 https://github.com/VanirAOSP/kernel_htc_msm8974/commit/62ec21519baa07a8a5b66c7f8b9b8c3bb337eeab