msm: qdsp6v2: Squash fixes
msm: audio: qdsp6v2: Fix for FBD TS inversion during AAC decoding

 -EOF flag of input buffer is not set properly.
 -set the flag of input buffer properly when meta info is enabled

Change-Id: I354fc16ad5bc8a9bcccdf150d1f138ce9744eeb0
CRs-Fixed: 431781
Signed-off-by: Preetam Singh Ranawat <apranawat@codeaurora.org>

Conflicts:
	arch/arm/mach-msm/qdsp6v2/audio_utils_aio.c

ASoC: msm: qdsp6v2: Ignore unexpected events from dsp
If dsp responds with an unexpected event, audio driver
should ignore the response and continue. Change removes
the BUG_ON and handles error gracefully.

Change-Id: Ie7bbf1ea9b47630dae38e89887e53883f8f1dc12
CRs-fixed: 512784
Signed-off-by: Gopikrishnaiah Anandan <agopik@codeaurora.org>

msm: audio: qdsp6v2: Fix for jumping to next track when seeked to end

- During tunnel playback, when seek to end of the clip, and then
  seek back to other position, music player jumps to next track.
- When progress bar is dragged to end of the clip or beginning,
  this will invoke fsync function. fsync function waits for WRITE_DONE
  event from DSP. AUDIO_STOP or AUDIO_FLUSH should wake up fsync,
  but this wake up is not happening, as result driver issuing EOS
  cmd to DSP. Causing music player to jump to next track.
- Fix is to wake up the fsync during AUDIO_STOP and AUDIO_FLUSH.

Change-Id: I1f6da57136b7a957861c4f731350c7fa4452f61b
Signed-off-by: Deepa Madiregama <dmadireg@codeaurora.org>

Conflicts:
	arch/arm/mach-msm/qdsp6v2/audio_utils_aio.c

ASoC: qcom: qdsp6v2: fix crash due to asynchronous read/write
Whenever an async read command is issued to dsp when
the driver is not enabled, there is a crash in userspace.
The read command is disallowed in the driver and an error
is propagated to userspace which results in the crash.
The driver is re-enabled in a subsequent asynchronous write
command anyway, so the async read command can be allowed.
Ensure that read is allowed by removing a superfluous check.

CRs-Fixed: 675131
Signed-off-by: Satya Krishna Pindiproli <satyak@codeaurora.org>
Change-Id: I5d0598e37d08ac1090f98c24be49fd99c132e22f

Based on commits:
  https://github.com/Team-SennyC2/senny_kernel-3.4/commit/ab2e72ed03bdb03eedf749801f3270a35809d834
  https://github.com/dexter93/kernel_htc_msm8660/commit/77954c5cfb9db0a2498ab8d96b3fd44159822f57
  https://github.com/dexter93/kernel_htc_msm8660/commit/567fe8ffbb133f0caf691179c90ad376d16b5366
  https://github.com/dexter93/kernel_htc_msm8660/commit/7b22cddd70d53353574c3f5ba023061716e60f19
  https://github.com/dexter93/kernel_htc_msm8660/commit/43dfbe078e4398d304d7121e8651874bb1e1973a